float to int

+++ Neelima_Stage3.py
@@ -1,7 +1,7 @@
 student_name = input("Enter the student's name: ")
-mark1 = float(input("Enter the first mark: "))
-mark2 = float(input("Enter the second mark: "))     
-mark3 = float(input("Enter the third mark: "))  
+mark1 = int(input("Enter the first mark: "))
+mark2 = int(input("Enter the second mark: "))     
+mark3 = int(input("Enter the third mark: "))  
 
 total_marks = mark1 + mark2 + mark3     
 percentage = (total_marks / 300) * 100
@@ -16,5 +16,3 @@ elif percentage >= 40:
     print(f"{student_name} has achieved a C grade.")    
 else:
     print(f"{student_name} has achieved a F grade.")
-    
-
